rename the select.py to mitigate the confliction with system module select
